requirements.txt
django==4.2.5
python_dotenv==1.0.0
psycopg2==2.9.9
django-autocomplete-light==3.9.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
